support very large files

+++ src/petro/box/hmhd.hpp
@@ -26,7 +26,7 @@ namespace box
             full_box(hmhd::TYPE, parent)
         { }
 
-        void read(uint32_t size, byte_stream& bs)
+        void read(uint64_t size, byte_stream& bs)
         {
             full_box::read(size, bs);
             m_max_pdu_size = bs.read_uint16_t();
