+++ src/petro/box/stdp.hpp
@@ -26,7 +26,7 @@ namespace box
             box(stdp::TYPE, parent)
         { }
 
-        void read(uint32_t size, byte_stream& bs)
+        void read(uint64_t size, byte_stream& bs)
         {
             box::read(size, bs);
             bs.skip(m_remaining_bytes);
