+++ src/petro/box/stsz.hpp
@@ -29,7 +29,7 @@ namespace box
             full_box(stsz::TYPE, parent)
         { }
 
-        void read(uint32_t size, byte_stream& bs)
+        void read(uint64_t size, byte_stream& bs)
         {
             full_box::read(size, bs);
             m_sample_size = bs.read_uint32_t();
