+++ src/petro/box/tkhd.hpp
@@ -27,7 +27,7 @@ namespace box
             full_box(tkhd::TYPE, parent)
         { }
 
-        void read(uint32_t size, byte_stream& bs)
+        void read(uint64_t size, byte_stream& bs)
         {
             full_box::read(size, bs);
